Change the design of the customized key setup
Make it completely compatible with the original setup for the special
case of a 256-octet key, and make it nearly compatible for the case of
keys larger than that.

Only perform actual custom processing for keys shorter than 256 octets,
and simplify this a lot compared to the previous base-128 encoding
approach by using single-octet base-256 encoding.