Update branch with UUID 9slawn4w20gvlddf89md6ak1f
Updated branch name: arcfour_xa_etm

Change the key setup considerably

Also drop the "drop3072" from the name.

First of all, we changed the number. Second, it does not matter because
we changed so much that the number of warmup cycles is just one of the
details what has been changed.

Not important enough any longer to be part of the project name.

Also addes a Q&A regarding design choices made in the modified key
setup.

The "etm" in the updated branch name shall reflect that fact that this
utility does not only encrypt, but also provides message authentication
via a MAC (message authentication code) which will be applied to the
encrypted message ("etm" stands for "encrypt then MAC").