Add ascii armor layer
Instead of using base-32 encoding by default for the output of certain
operations such as hash calculations, always default to binary output.

The -a option can then be added to generate ASCII-armored output instead

For instance,

$ a4x -H < file | hexdump -C

will generate a hash digest in binary form which needs to be hex-dumped
for display, where

$ a4x -Ha < file

will output an ASCII-armored message digest which can be displayed.