Replace base-32 encoding with base-64 encoding
When base-32 ascii-armoring was restricted to encoding hash values and
MACs, it was OK to use base-32 because the strings where still short.

When generating ascii-armored encrypted text, however, encoding
efficiency becomes more important, making base-32 less attractive.

The only question remaining is whether the -a/-s options are required at
all any longer, because the combination of "base64", "fold" and "tr" can
achieve the same effects.

However, "base64" has not been standardized by POSIX, and as a
consequence there are many incompatible implementations. It seems
therefore prudent not to assume any of them must be available, and
incorporate a base-64 codec in the application.