Remove unnecessary backslashes from Makefile rules
Before there was a script for printing out the dependencies, this had to
be done manually. Therefore a redundant backslash was added to the last
line of a multiline dependency rule in order to allow filtering the
dependency block with the "sort" command from inside the editor.

This is no longer necessary, so the trailing backslash serves no
purpose. Update the dependency rule generator script in order to remove
it.